requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.36
psycopg2-binary==2.9.9
alembic==1.13.2
python-jose==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.8.2
pydantic-settings==2.4.0
python-multipart==0.0.9
tenacity==9.0.0
PyJWT
email-validator==2.2.0

